requirements.txt
--index-url https://download.pytorch.org/whl/cu126
--extra-index-url https://pypi.org/simple

torch>=2.7
torchvision>=0.22
numpy>=2.0
Pillow>=10.0
einops>=0.8
scikit-learn>=1.5
pycocotools>=2.0
tqdm>=4.60
termcolor>=1.1.0
torchmetrics>=0.11
matplotlib
